requirements.txt
av==13.1.0
depthai==2.28.0.0
numpy==2.1.3
pillow==11.0.0
tornado==6.4.1
